refactor(data): 통합된 DefaultVoiceNoteRepository 구현
