--- Data/Sources/Repositories/VoiceNotes/VoiceNoteFetchRepositoryError+Mapping.swift
@@ -1,14 +1,0 @@
-import Domain
-
-extension VoiceNoteFetchRepositoryError {
-    init(_ error: CoreDataStorageError) {
-        switch error {
-        case .fetchFailed:
-            self = .unknown(error)
-        case .fetchAllFailed:
-            self = .unknown(error)
-        default:
-            self = .unknown(error)
-        }
-    }
-}
